% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{bisection for root finding. Vectorized.}
\usage{
bisection(
  f,
  lower,
  upper,
  ...,
  maxiter = 1000,
  tol = 1e-08,
  search_step = 1,
  auto_adjust_interval = TRUE
)
}
\arguments{
\item{f}{If auto_adjust_interval is TRUE, f must be an increasing function.}

\item{auto_adjust_interval}{if signs of f(lower) and f(upper) are not opposite, search for new interval. If TRUE ,f must be an increasing function}
}
\description{
modified from https://stat.ethz.ch/pipermail/r-help/2012-November/340295.html
}
