% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_normal.R
\name{ebpm_normal}
\alias{ebpm_normal}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
ebpm_normal(
  x,
  s = NULL,
  g_init = NULL,
  fix_g = FALSE,
  q_init = NULL,
  maxiter = 1000,
  tol = 1e-05,
  vga_tol = 1e-05,
  conv_type = "elbo"
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}

\item{prior_mean}{prior mean}

\item{prior_var}{prior variance}
}
\value{
a list of
 \item{posterior:}{mean_log/var_log is the posterior mean/var of mu, mean is the posterior of exp(mu)}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
}
\description{
Gaussian prior, Gaussian posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(\beta,\sigma^2).}
}
\examples{
\dontrun{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
ebpm_normal(x)
}
}
