% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d1_mu}
\alias{f_nm_d1_mu}
\title{gmm likelihood derivative wrt prior mean mu}
\usage{
f_nm_d1_mu(x, s, w, mu, grid)
}
\value{
gradient w.r.t prior mean, a scalar
}
\description{
gmm likelihood derivative wrt prior mean mu
}
