% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d1_s2}
\alias{f_nm_d1_s2}
\title{gmm likelihood derivative wrt s2}
\usage{
f_nm_d1_s2(x, s, w, mu, grid)
}
\value{
a vector of derivative df_nm/ds2, length n
}
\description{
gmm likelihood derivative wrt s2
}
