% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d1_z}
\alias{f_nm_d1_z}
\title{gmm likelihood derivative wrt z}
\usage{
f_nm_d1_z(x, s, w, mu, grid)
}
\value{
a vector of gradient df/dz, length n
}
\description{
gmm likelihood derivative wrt z
}
