% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d2_z}
\alias{f_nm_d2_z}
\title{gmm likelihood 2nd derivative wrt z}
\usage{
f_nm_d2_z(x, s, w, mu, grid)
}
\value{
a vector of second derivative d^2f/dz^2, length n
}
\description{
gmm likelihood 2nd derivative wrt z
}
