% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d2_zs2}
\alias{f_nm_d2_zs2}
\title{gmm likelihood derivative wrt z,s2}
\usage{
f_nm_d2_zs2(x, s, w, mu, grid)
}
\value{
a vector of second derivative d^2f_nm/dzds2, length n
}
\description{
gmm likelihood derivative wrt z,s2
}
