% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d3_z}
\alias{f_nm_d3_z}
\title{gmm likelihood 3rd derivative wrt z}
\usage{
f_nm_d3_z(x, s, w, mu, grid)
}
\value{
a vector of third derivative d^3f/dz^3, length n
}
\description{
gmm likelihood 3rd derivative wrt z
}
