% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_inversion.R
\name{f_obj}
\alias{f_obj}
\title{objective function for inversion method}
\usage{
f_obj(params, y, grid)
}
\arguments{
\item{params}{(theta,w,mu)}
}
\description{
objective function for inversion method
}
