% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_penalized_compound.R
\name{f_obj_ebnm_compound}
\alias{f_obj_ebnm_compound}
\title{objective function}
\usage{
f_obj_ebnm_compound(theta, y, s, grid)
}
\arguments{
\item{theta}{(mu_bar,w)}

\item{grid}{prior sds}
}
\description{
objective function
}
