% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_penalized_inversion.R
\name{f_obj_grad_ebnm_inversion}
\alias{f_obj_grad_ebnm_inversion}
\title{objective function gradient ebnm_compound}
\usage{
f_obj_grad_ebnm_inversion(params, y, s, grid)
}
\arguments{
\item{grid}{prior sds}

\item{theta}{(theta,w)}
}
\description{
objective function gradient ebnm_compound
}
