% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{gen_data_exp}
\alias{gen_data_exp}
\title{generate simulation data}
\usage{
gen_data_exp(
  n = 1000,
  n_simu = 100,
  w = 0.8,
  prior_mean = 2,
  prior_var0 = 0,
  prior_var1 = 3,
  seed = 12345
)
}
\description{
generate simulation data
}
