% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{gen_data_gamma}
\alias{gen_data_gamma}
\title{generate simulation data, gamma}
\usage{
gen_data_gamma(
  n = 1000,
  n_simu = 100,
  prior = "gamma",
  shape0 = 2,
  rate0 = 2,
  shape1 = 10,
  rate1 = 2,
  w = 0.8,
  seed = 12345
)
}
\description{
generate simulation data, gamma
}
\details{
The DGP is
\deqn{x_i\sim Poisson(\lambda_i),}
\deqn{\lambda_i\sim w\times Gamma(shape0,rate0) + (1-w)\times Gamma(shape1,rate1).}
}
