% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{gen_data_identity}
\alias{gen_data_identity}
\title{generate simulation data}
\usage{
gen_data_identity(
  n = 1000,
  n_simu = 100,
  prior = "gamma",
  gamma_shape = 10,
  gamma_rate = 2,
  t_df = 3,
  t_delta = 2,
  quasi_sparse = 0.1,
  exp_rate = 0.5,
  w = 0.8,
  seed = 12345
)
}
\description{
generate simulation data
}
