% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{gen_data_log_link}
\alias{gen_data_log_link}
\title{generate simulation data, log-link}
\usage{
gen_data_log_link(
  n = 1000,
  n_simu = 100,
  w = 0.8,
  prior_mean = 2,
  prior_var0 = 0,
  prior_var1 = 2,
  seed = 12345
)
}
\description{
generate simulation data, log-link
}
\details{
The DGP is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim w\times N(m,\sigma^2_0) + (1-w)\times N(m,\sigma^2_1).}
}
