% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{gen_data_t}
\alias{gen_data_t}
\title{generate simulation data, t}
\usage{
gen_data_t(
  n = 1000,
  n_simu = 100,
  prior = "t",
  t_df = 3,
  t_delta = 2,
  w = 0.8,
  seed = 12345
)
}
\description{
generate simulation data, t
}
\details{
The DGP is
\deqn{x_i\sim Poisson(\lambda_i),}
\deqn{\lambda_i\sim w\times \delta_{c} + (1-w)\times (|t_{df}|+delta_{2c}).}
}
