% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_cstr}
\alias{h_cstr}
\title{calculate constraint function}
\usage{
h_cstr(params, y, grid)
}
\arguments{
\item{theta}{(z,s2,a)}
}
\value{
a vector of length n, constraint function values
}
\description{
calculate constraint function
}
