% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_cstr_d1_g_calc}
\alias{h_cstr_d1_g_calc}
\title{The gradient of constraint function w.r.t a}
\usage{
h_cstr_d1_g_calc(z, s, a, mu, grid)
}
\value{
a n*K matrix
}
\description{
The gradient of constraint function w.r.t a
}
