% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_cstr_d1_s2_calc}
\alias{h_cstr_d1_s2_calc}
\title{calculate constraint function derivative wrt s2}
\usage{
h_cstr_d1_s2_calc(z, s, w, mu, grid)
}
\arguments{
\item{theta}{(z,s2,a,mu)}
}
\value{
a vector of length n gradient
}
\description{
calculate constraint function derivative wrt s2
}
