% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_obj}
\alias{h_obj}
\title{objective function, sum over h_i()}
\usage{
h_obj(params, y, grid)
}
\arguments{
\item{params}{(z,s2,a,mu)}

\item{y}{data vector}

\item{grid}{prior sds}
}
\value{
objective value, a scalar
}
\description{
objective function, sum over h_i()
}
