% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_obj_d1_g_calc}
\alias{h_obj_d1_g_calc}
\title{objective function derivative wrt a}
\usage{
h_obj_d1_g_calc(z, s, a, mu, y, grid)
}
\arguments{
\item{y}{data vector}

\item{grid}{prior sds}

\item{theta}{(z,s2,a)}
}
\value{
a length K vector of gradients
}
\description{
objective function derivative wrt a
}
