% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_obj_d1_s2_calc}
\alias{h_obj_d1_s2_calc}
\title{objective function derivative wrt s2}
\usage{
h_obj_d1_s2_calc(z, s, w, mu, y, grid)
}
\arguments{
\item{y}{data vector}

\item{grid}{prior sds}

\item{theta}{(z,s2,a)}
}
\value{
a length n vector of gradients
}
\description{
objective function derivative wrt s2
}
