% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{h_obj_d1_z_calc}
\alias{h_obj_d1_z_calc}
\title{objective function derivative wrt z}
\usage{
h_obj_d1_z_calc(z, s, w, mu, y, grid)
}
\arguments{
\item{y}{data vector}

\item{grid}{prior sds}

\item{params}{(z,s2,a,mu)}
}
\value{
a length n vector of gradients
}
\description{
objective function derivative wrt z
}
