% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm}
\alias{l_nm}
\title{log marginal likelihood of normal mean model}
\usage{
l_nm(x, s, w, mu, grid)
}
\arguments{
\item{x}{data vector of length n}

\item{s}{standard error}

\item{w}{prior weights}

\item{mu}{prior mean}

\item{grid}{grid of sd in prior}
}
\value{
a vector of length n
}
\description{
log marginal likelihood of normal mean model
}
