% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d1_a}
\alias{l_nm_d1_a}
\title{gmm log-likelihood derivative wrt a}
\usage{
l_nm_d1_a(x, s, a, mu, grid)
}
\value{
a matrix of gradient, size n*K
}
\description{
gmm log-likelihood derivative wrt a
}
