% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d1_g}
\alias{l_nm_d1_g}
\title{gmm log-lik derivative of prior}
\usage{
l_nm_d1_g(z, theta, s, a, mu, grid)
}
\value{
derivative of l_nm(z(theta);g,s^2(theta)) w.r.t prior (a,mu)
}
\description{
gmm log-lik derivative of prior
}
