% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d1_s2}
\alias{l_nm_d1_s2}
\title{gmm log-likelihood derivative wrt s2}
\usage{
l_nm_d1_s2(x, s, w, mu, grid)
}
\value{
a vector of derivative dl_nm/ds2, length n
}
\description{
gmm log-likelihood derivative wrt s2
}
