% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d1_theta}
\alias{l_nm_d1_theta}
\title{gmm log-likelihood derivative wrt theta, the posterior mean}
\usage{
l_nm_d1_theta(z, theta, s, w, mu, grid)
}
\value{
derivative of l_nm(z(theta);g,s^2(theta)) w.r.t theta
}
\description{
gmm log-likelihood derivative wrt theta, the posterior mean
}
