% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d1_z}
\alias{l_nm_d1_z}
\title{gmm log-likelihood derivative wrt z}
\usage{
l_nm_d1_z(x, s, w, mu, grid)
}
\value{
a vector of derivative dl_nm/dz, length n
}
\description{
gmm log-likelihood derivative wrt z
}
