% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d2_z}
\alias{l_nm_d2_z}
\title{gmm log-likelihood 2nd derivative wrt z}
\usage{
l_nm_d2_z(x, s, w, mu, grid)
}
\value{
a vector of second derivative d^2l_nm/dz^2, length n
}
\description{
gmm log-likelihood 2nd derivative wrt z
}
