% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d2_za}
\alias{l_nm_d2_za}
\title{gmm log-likelihood derivative wrt z,a}
\usage{
l_nm_d2_za(x, s, a, mu, grid)
}
\value{
a matrix of gradient, size n*K
}
\description{
gmm log-likelihood derivative wrt z,a
}
