% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d2_zs2}
\alias{l_nm_d2_zs2}
\title{gmm log-likelihood derivative wrt z,s2}
\usage{
l_nm_d2_zs2(x, s, w, mu, grid)
}
\value{
a vector of second derivative d^2l_nm/dzds2, length n
}
\description{
gmm log-likelihood derivative wrt z,s2
}
