% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{l_nm_d3_z}
\alias{l_nm_d3_z}
\title{gmm log-likelihood 3rd derivative wrt z}
\usage{
l_nm_d3_z(x, s, w, mu, grid)
}
\value{
a vector of third derivative d^3l_nm/dz^3, length n
}
\description{
gmm log-likelihood 3rd derivative wrt z
}
