% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{nm_density}
\alias{nm_density}
\title{calc normal density of x, mixture}
\usage{
nm_density(x, s, mu, grid)
}
\value{
a n by K matrix of normal density
}
\description{
calc normal density of x, mixture
}
