% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{nm_penalty_compound_grad}
\alias{nm_penalty_compound_grad}
\title{gradient of compound penalty function}
\usage{
nm_penalty_compound_grad(theta, s, w, mu, grid)
}
\description{
-l'(theta) - s^2l'(theta)l''(theta)
}
