% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{nm_penalty_compound_hess}
\alias{nm_penalty_compound_hess}
\title{hessian of compound penalty function}
\usage{
nm_penalty_compound_hess(theta, s, w, mu, grid)
}
\description{
-l''(theta) - s^2(l''(theta)^2+l'''(theta)l'(theta))
}
