% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_GG.R
\name{pois_mean_GG}
\alias{pois_mean_GG}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
pois_mean_GG(
  x,
  s = NULL,
  prior_mean = NULL,
  prior_var = NULL,
  optim_method = "L-BFGS-B",
  maxiter = 1000,
  tol = 1e-05,
  m_init = NULL,
  v_init = NULL
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{prior_mean}{prior mean}

\item{prior_var}{prior variance}

\item{optim_method}{optimization method in `optim` function}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}
}
\value{
a list of
 \item{posterior:}{posteriorMean_latent is the posterior mean of mu, posteriorMean_mean is the posterior of exp(mu)}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
}
\description{
Gaussian prior, Gaussian posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(\beta,\sigma^2).}
}
\examples{
\dontrun{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_GG(x)
}
}
