% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vebpm_VGA_optimize.R
\name{pois_mean_GG_opt_obj_gradient0}
\alias{pois_mean_GG_opt_obj_gradient0}
\title{calculate VGA Poisson objective function gradient}
\usage{
pois_mean_GG_opt_obj_gradient0(theta, x, s, beta, sigma2, n)
}
\arguments{
\item{theta}{(m,v)}
}
\description{
calculate VGA Poisson objective function gradient
}
