% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vebpm_VGA_optimize.R
\name{pois_mean_GG_opt_obj_m_gradient}
\alias{pois_mean_GG_opt_obj_m_gradient}
\title{calculate gradient of VGA Poisson objective function of posterior mean}
\usage{
pois_mean_GG_opt_obj_m_gradient(m, x, s, beta, sigma2)
}
\description{
calculate gradient of VGA Poisson objective function of posterior mean
}
