% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_GMG.R
\name{pois_mean_GMG}
\alias{pois_mean_GMG}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
pois_mean_GMG(
  x,
  s = NULL,
  w = NULL,
  prior_mean = NULL,
  mixsd = NULL,
  optim_method = "L-BFGS-B",
  maxiter = 1000,
  tol = 1e-05
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{w}{prior weights}

\item{prior_mean}{prior mean}

\item{mixsd}{prior sd grids}

\item{optim_method}{optimization method in `optim` function}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}
}
\value{
a list of
 \item{posteriorMean:}{posterior mean}
 \item{posteriorVar:}{posterior var}
 \item{obj_value:}{objective function values}
 \item{priorMean:}{prior mean}
 \item{sigma2:}{prior variance}
}
\description{
Gaussian Mixture prior, Gaussian posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim \sum_k N(\beta,\sigma_k^2).}
}
\examples{
n = 10000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_GMG(x)
}
