% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_GMG.R
\name{pois_mean_GMG_opt_obj_gradient}
\alias{pois_mean_GMG_opt_obj_gradient}
\title{calculate gradient vector for optim}
\usage{
pois_mean_GMG_opt_obj_gradient(theta, x, s, qz, beta, sigma2k, n)
}
\description{
calculate gradient vector for optim
}
