% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_GMGM.R
\name{pois_mean_GMGM}
\alias{pois_mean_GMGM}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
pois_mean_GMGM(
  x,
  s = NULL,
  w = NULL,
  prior_mean = NULL,
  mixsd = NULL,
  point_mass = TRUE,
  optim_method = "L-BFGS-B",
  maxiter = 1000,
  tol = 1e-05
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{w}{prior weights}

\item{prior_mean}{prior mean}

\item{mixsd}{prior sd}

\item{point_mass}{whether put a point-mass in the prior}

\item{optim_method}{optimization method in `optim` function}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}
}
\value{
a list of
 \item{posterior:}{posteriorMean_latent is the posterior mean of mu, posteriorMean_mean is the posterior of exp(mu)}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
}
\description{
Gaussian Mixture prior, Gaussian Mixture posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim \sum_k w_k N(\beta,\sigma_k^2).}
}
\examples{
\dontrun{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_GMGM(x)
}
}
