% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_log1exp.R
\name{pois_mean_log1exp}
\alias{pois_mean_log1exp}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
pois_mean_log1exp(
  x,
  ebnm_params = NULL,
  tol = 1e-05,
  maxiter = 1000,
  kapa = NULL
)
}
\arguments{
\item{x}{data vector}

\item{ebnm_params}{a list of `ebnm` parameters}

\item{tol}{tolerance for stopping the updates}

\item{maxiter}{max number of iterations}

\item{kapa}{default to be 0.25 + 0.17*max(x)}
}
\value{
a list of
 \item{posterior:}{posteriorMean_latent is the posterior mean of mu, posteriorMean_mean is for Poisson mean}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
 \item{fit:}{fitted object from `ebnm`}
}
\description{
Poisson mean problem, log(1+exp(x)) link function.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\log(1+\exp(\mu_i))),}
\deqn{\mu_i\sim g(\cdot).}
}
\examples{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_log1exp(x)
}
