% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_penalized_compound.R
\name{pois_mean_penalized_compound}
\alias{pois_mean_penalized_compound}
\title{Solve poisson penalized problem via compound method}
\usage{
pois_mean_penalized_compound(
  x,
  w = NULL,
  prior_mean = NULL,
  mixsd = NULL,
  point_mass = TRUE,
  maxiter = 100,
  verbose = FALSE,
  tol = 1e-04
)
}
\arguments{
\item{x}{data vector}

\item{w}{prior weights}

\item{prior_mean}{prior mean}

\item{mixsd}{prior sd}

\item{maxiter}{max number of iterations}
}
\value{
a list of
 \item{posterior:}{posteriorMean_latent is the posterior mean of mu, posteriorMean_mean is the posterior of exp(mu)}
 \item{fitted_g:}{estimated prior}
}
\description{
Formulate Poisson mean problem as likelihood + penalty
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim \sum_k w_k N(\beta,\sigma_k^2).}
}
\examples{
n = 10000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_penalized_compound(x)
}
