% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_spliting2.R
\name{pois_mean_split}
\alias{pois_mean_split}
\title{Solve Poisson mean problem by a splitting method}
\usage{
pois_mean_split(
  x,
  s = NULL,
  sigma2 = NULL,
  est_sigma2 = TRUE,
  tol = 1e-05,
  maxiter = 1000,
  ebnm_params = NULL,
  mu_pm_init = NULL
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{sigma2}{prior variance}

\item{est_sigma2}{whether fix sigma2 at input or update it in iterations}

\item{tol}{tolerance for stopping the updates}

\item{maxiter}{max number of iterations}

\item{ebnm_params}{a list of `ebnm` parameters}

\item{mu_pm_init}{init value of E_q(mu), default to log(1+x)}
}
\value{
a list of
 \item{posterior:}{posterior mean of mu, posterior of exp(mu), etc.}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
}
\description{
Solve Poisson mean problem by a splitting method
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(b_i,\sigma^2),}
\deqn{b_i\sim g(.).}
}
\examples{
\dontrun{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_split(x)
}
}
