% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mean_spliting.R
\name{pois_mean_split_init_b}
\alias{pois_mean_split_init_b}
\title{Solve Poisson mean problem by a splitting method}
\usage{
pois_mean_split_init_b(
  x,
  s = NULL,
  sigma2 = NULL,
  est_sigma2 = TRUE,
  tol = 1e-05,
  maxiter = 1000,
  ebnm_params = NULL,
  optim_method = "L-BFGS-B",
  b_pm_init = NULL
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vecto}

\item{sigma2}{prior variance}

\item{est_sigma2}{whether fix sigma2 at input or update it in iterations}

\item{tol}{tolerance for stopping the updates}

\item{maxiter}{max number of iterations}

\item{ebnm_params}{a list of `ebnm` parameters}

\item{optim_method}{optimization method in `optim` function}
}
\value{
a list of
 \item{posterior:}{posteriorMean_latent is the posterior mean of mu, posteriorMean_mean is the posterior of exp(mu)}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
}
\description{
Solve Poisson mean problem by a splitting method
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(b_i,\sigma^2),}
\deqn{\b_i\sim g(.).}
}
\examples{
\dontrun{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
pois_mean_split(x)
}
}
