% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{S_inv_obj_jac}
\alias{S_inv_obj_jac}
\title{jacobian of  objective function for solving S(z)=theta}
\usage{
S_inv_obj_jac(z, theta, s, w, mu, grid)
}
\description{
jacobian of  objective function for solving S(z)=theta
}
