% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{simu_study_poisson_mean}
\alias{simu_study_poisson_mean}
\title{run the algorithms}
\usage{
simu_study_poisson_mean(
  sim_data,
  ebnm_params = list(prior_family = "normal_scale_mixture"),
  tol = 1e-05,
  maxiter = 2000,
  n_cores = 10
)
}
\description{
run the algorithms
}
