% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vebpm_VGA_optimize.R
\name{vga_optimize}
\alias{vga_optimize}
\title{optimize vga for posterior mean and var}
\usage{
vga_optimize(init_val, x, s, beta, sigma2, method = "lbfgs")
}
\description{
optimize vga for posterior mean and var
}
