% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vebpm_VGA_optimize.R
\name{vga_optimize_m}
\alias{vga_optimize_m}
\title{optimize vga for posterior mean then set v = v(m)}
\usage{
vga_optimize_m(init_val, x, s, beta, sigma2, method = "lbfgs")
}
\description{
optimize vga for posterior mean then set v = v(m)
}
