% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vga_pois_solver.R
\name{vga_pois_solver}
\alias{vga_pois_solver}
\title{Optimize vga poisson, using bisection for v}
\usage{
vga_pois_solver(init_val, x, s, beta, sigma2, maxiter = 1000, tol = 1e-08)
}
\arguments{
\item{x, s}{data and scale factor}

\item{beta, sigma2}{prior mean and variance}
}
\description{
Optimize vga poisson, using bisection for v
}
