% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{z_d1_theta}
\alias{z_d1_theta}
\title{derivative z wrt theta when z (and s2) is a function of theta.}
\usage{
z_d1_theta(z, theta, s, w, mu, grid)
}
\description{
derivative z wrt theta when z (and s2) is a function of theta.
}
