% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{bisection}
\alias{bisection}
\title{bisection for root finding. Vectorized}
\usage{
bisection(f, lower, upper, ..., maxiter = 100, tol = 1e-08)
}
\description{
from https://stat.ethz.ch/pipermail/r-help/2012-November/340295.html
}
