% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{f_nm_d2_zmu}
\alias{f_nm_d2_zmu}
\title{gmm likelihood derivative wrt z,mu}
\usage{
f_nm_d2_zmu(x, s, w, mu, grid)
}
\value{
a vector gradient
}
\description{
gmm likelihood derivative wrt z,mu
}
