% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{S}
\alias{S}
\title{#'posterior mean operator
S = function(x,s,w,mu,grid){
  K = length(w)
  g = normalmix(pi=w,mean=rep(mu,K),sd=grid)
  fit.ash = ashr::ash(x,s,g=g,fixg=T)
  fit.ash$result$PosteriorMean
}}
\usage{
S(x, s, w, mu, grid)
}
\description{
#'posterior mean operator
S2 = function(x,s,w,mu,grid){
  lW = matrix(log(w),nrow=length(x),ncol=length(grid),byrow=T)
  pw = lW+dnorm(x,mean=mu,sd=outer(s^2,grid^2,FUN='+'),log=TRUE)
  pw = pw - apply(pw,1,max)
  pw = exp(pw)/rowSums(exp(pw))
  temp  = outer(s^2,grid^2,FUN='/')
  pm = x/(1+temp) + mu/(1+1/temp)
  return(rowSums(pw*pm))
}
posterior mean operator
}
