% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mean_model_utils.R
\name{S_inv}
\alias{S_inv}
\title{inverse PM operator}
\usage{
S_inv(theta, s, w, mu, grid)
}
\description{
#' jacobian of  objective function for solving S(z)=theta
S_inv_obj_jac = function(z,theta,s,w,mu,grid){
  return(diag(c(1+s^2*l_nm_d2_z(z,s,w,mu,grid))))
}
}
