% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{simu_study_poisson_mean}
\alias{simu_study_poisson_mean}
\title{run the algorithms}
\usage{
simu_study_poisson_mean(
  sim_data,
  ebnm_params = list(prior_family = "normal_scale_mixture"),
  tol = 1e-08,
  maxiter = 2000,
  n_cores = 10,
  method_list = c("GG", "GMG", "GMGM", "GMGM_pointmass", "nb_lb", "nb_pg", "log1exp",
    "split", "split_mixture", "penalty_compound", "penalty_inversion",
    "ash_pois_identity", "ash_pois_log", "ebpm_gamma", "ebpm_exp_mixture"),
  save_data = TRUE
)
}
\description{
run the algorithms
}
